+++ lib/table_sync/orm_adapter/sequel.rb
@@ -2,10 +2,6 @@
 
 module TableSync::ORMAdapter
   class Sequel < Base
-    def primary_key
-      object.pk_hash
-    end
-
     def attributes
       object.values
     end