+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 ## [6.0.0] - 2021-10-01
 ### Added
 
-A lot of specs for all the refactoring.
+- A lot of specs for all the refactoring.
+- Docs
+- 100% coverage
 
 ### Changed
 - Heavy refactoring of Publisher and BatchPublisher.
@@ -36,6 +38,10 @@ Now custom attributes for destruction will be called on instances.
 - Obj.table_sync_destroy_attributes() -> Obj#attributes_for_destroy
 
 8. Event constants are now kept in one place.
+
+### Removed
+
+- Plugin Errors
 
 ## [5.1.0] - 2021-09-09